+++ Projects/App/Sources/Navigation/LoginNavigation.swift
@@ -9,7 +9,6 @@ import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 struct LoginNavigation: View {
     @ObservedObject var vm: NotLoggedViewModel
