+++ Projects/App/Sources/UgglanStore.swift
@@ -4,7 +4,6 @@ import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 struct UgglanState: StateProtocol {
     var isDemoMode: Bool = false
