+++ Projects/Authentication/Sources/SEBankID/BankIDLoginQRView.swift
@@ -2,7 +2,6 @@ import Combine
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 public struct BankIDLoginQRView: View {
     @StateObject var vm = BankIDViewModel()
