+++ Projects/hCoreUI/Sources/Views/CoverageView.swift
@@ -1,6 +1,5 @@
 import SwiftUI
 import hCore
-import hGraphQL
 
 public struct CoverageView: View {
     let limits: [InsurableLimits]
