+++ Projects/hCoreUI/Sources/Views/UpdateOSScreen.swift
@@ -1,6 +1,5 @@
 import SwiftUI
 import hCore
-import hGraphQL
 
 public struct UpdateOSScreen: View {
